Test Environment/cloud-resources/development_support/validate/requirements.txt
pyyaml==5.3
cerberus==1.3.2
Jinja2==2.10.3